Replace loader function with protected route component
There is still problem when reloading nested protected route that loads ./ route. By setting to Login component instead of null for ProtectedRoute it will render login component but url will not be replaced which is ok for now but it would be nice if this could in future be solved. Reloading nested route for now works.